requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.23
langgraph>=0.0.26
langchain-groq>=0.1.0
langsmith>=0.1.85

# Database
sqlalchemy==2.0.23
alembic==1.13.0
pandas==2.1.3
sqlite-utils==3.35.2

# Vector Store
chromadb==0.4.18
sentence-transformers==2.3.1
huggingface-hub>=0.20.0

# LLM & Embeddings
groq==0.4.1
openai==1.6.1

# HTTP & WebSocket
httpx==0.25.2
websockets==12.0
python-multipart==0.0.6

# Utilities
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
beautifulsoup4==4.12.2
duckduckgo-search==3.9.6

# Visualization
matplotlib==3.8.2
plotly==5.18.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

# Development
black==23.12.0
flake8==6.1.0
mypy==1.7.1
